Time: 193 ms (42.63%), Space: 36.8 MB (46.92%) - LeetHub

+++ 1383-maximum-performance-of-a-team/1383-maximum-performance-of-a-team.cpp
@@ -1,6 +1,6 @@
 class Solution { 
     private:
-    int mod=1e9+7;
+    long mod=1e9+7;
 public:
     int maxPerformance(int n, vector<int>& s, vector<int>& e, int k) {
         vector<pair<int,int>>ds;
